Fix tab dot pulsing for threads that no longer render
The TabBar's per-tab dot counted comment threads with a naive
quote-includes check, while the comment gutter renders a thread only when
the overlay's resolver attaches it — and that resolver validates quote
matches against the anchor's stored surrounding context. Once a commented
passage was deleted or rewritten, the gutter correctly detached the
thread, but the dot kept pulsing as long as the same string appeared
anywhere else in the document — near-permanent in LaTeX files, where
boilerplate like \item and \begin{itemize} recurs constantly. The result
was a pulsing dot with no visible thread to review.

Extract the overlay's quote-matching ladder (occurrence-index lookup,
multi-line first-line fallback, context validation) into a shared
matchCommentAnchor in ydoc-codec, and use it from both the overlay's
fallback resolver and the per-tab thread counting behind the dots. One
implementation means the two surfaces can't drift apart again — the drift
happened when the overlay gained its context check and the dot count
didn't.

Also stamp anchor context at creation in the provider comment_doc path
(matching the Claude-path comment tools) so those threads get the same
protection against re-attaching to unrelated occurrences, and replace
mcp-doc-tools' private nthOccurrenceIndex with the shared nthIndexOf.
